+++ client/src/main.rs
@@ -15,9 +15,11 @@ use avail_light_core::{
 	network::{
 		self,
 		p2p::{
-			self, configuration::LibP2PConfig, extract_block_num, forward_p2p_events,
-			init_and_start_p2p_client, p2p_restart_manager, OutputEvent as P2pEvent,
-			BOOTSTRAP_LIST_EMPTY_MESSAGE, MINIMUM_P2P_CLIENT_RESTART_INTERVAL,
+			self,
+			configuration::{AutoNatMode, LibP2PConfig},
+			extract_block_num, forward_p2p_events, init_and_start_p2p_client, p2p_restart_manager,
+			OutputEvent as P2pEvent, BOOTSTRAP_LIST_EMPTY_MESSAGE,
+			MINIMUM_P2P_CLIENT_RESTART_INTERVAL,
 		},
 		rpc::{self, OutputEvent as RpcEvent},
 		Network,
@@ -27,8 +29,8 @@ use avail_light_core::{
 	sync_finality::SyncFinality,
 	telemetry::{self, otlp::Metrics, MetricCounter, MetricValue, ATTRIBUTE_OPERATING_MODE},
 	types::{
-		load_or_init_suri, Delay, IdentityConfig, MaintenanceConfig, NetworkMode, PeerAddress,
-		SecretKey, Uuid,
+		load_or_init_suri, Delay, IdentityConfig, KademliaMode, MaintenanceConfig, NetworkMode,
+		PeerAddress, SecretKey, Uuid,
 	},
 	updater,
 	utils::{self, default_subscriber, install_panic_hooks, json_subscriber, spawn_in_span},
@@ -41,7 +43,10 @@ use color_eyre::{
 };
 use config::RuntimeConfig;
 use kate::couscous;
-use libp2p::kad::{Mode, QueryStats, RecordKey};
+use libp2p::{
+	kad::{Mode, QueryStats, RecordKey},
+	Multiaddr,
+};
 use std::{collections::HashMap, fs, path::Path, sync::Arc, time::Duration};
 use tokio::{
 	select,
@@ -66,6 +71,11 @@ static GLOBAL: Jemalloc = Jemalloc;
 mod cli;
 mod config;
 mod tracking;
+
+const EXTERNAL_ADDRESS_NOT_SET_MESSAGE: &str = r#"
+External address must be set if auto nat is disabled, and automatic server mode is enabled or kademlia mode is 'server'.
+Either use a '--external-address' parameter or add 'external_address' in the configuration file.
+"#;
 
 /// Light Client for Avail Blockchain
 async fn run(
@@ -546,6 +556,22 @@ pub fn load_runtime_config(opts: &CliOpts) -> Result<RuntimeConfig> {
 
 	if let Some(address_blacklist) = &opts.address_blacklist {
 		cfg.libp2p.address_blacklist = address_blacklist.clone();
+	}
+
+	if let Some(external_address) = &opts.external_address {
+		let multiaddr: Multiaddr = external_address.parse().map_err(|error| {
+			eyre!("Failed to parse external address {external_address}: {error}")
+		})?;
+		cfg.libp2p.external_address = Some(multiaddr);
+	}
+
+	if !cfg.libp2p.local_test_mode
+		&& cfg.libp2p.external_address.is_none()
+		&& cfg.libp2p.behaviour.auto_nat_mode != AutoNatMode::Client
+		&& (cfg.libp2p.kademlia.operation_mode == KademliaMode::Server
+			|| cfg.libp2p.kademlia.automatic_server_mode)
+	{
+		return Err(eyre!("{EXTERNAL_ADDRESS_NOT_SET_MESSAGE}"));
 	}
 
 	Ok(cfg)