+++ client/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.13.2]
 
+- Add --external-address CLI parameter for use on servers without autonat
 - Update `avail-light-core` to 1.2.11
 
 ## [1.13.1](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.13.1) - 2025-09-08